moved QR decomposition to a separate header file
